backend/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
paho-mqtt==1.5.1
psycopg2-binary==2.9.1
pymongo==3.11.4
six==1.16.0
Werkzeug==2.0.1
